inventario-personal/backend/auth_service/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
sqlalchemy==1.4.23
mysql-connector-python==8.0.27
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4